Coordinates now start as Vectors, random selection having issues
Have made the selection of coordinates of both Reaper and Ghost Leviathans now start as Vector3, as opposed to float arrays with all 3 coordinates inside.
Have made a class, GhostCoord, in order to accomodate this change whilst still keeping the 'Adult or Juvenile' value.
Finally, have started work on instead randomly selecting from the coord array and adding that to the save file. However, the more I do this, the more I suspect the coord array should be a list. Next commit will likely be just that.