Started final implementation of save and loading
First commit towards the new save and load system Eldritch suggested; check if the coord file is null on starting load, populate it with new coordinates if so, leave it alone if not. Spawn in the leviathans from the Coord File on player.Awake event.

Have also cleaned up the code a bit by using 'Vector3' instead of 'UnityEngine.Vector3'.