Added skip in code if spawn intensity is 0
Populating the coord file skips if both spawn intensities are set to 0.

Accompanying this, most of the functions of the mod do not occur if they detect the coord file is null.
This means a user can set the intensity to 0 and safely play a save without worrying about spawning leviathans, but can later choose to by restarting and setting the intensity higher.