Updating coord format works
Updating the format of both ReaperCoords and GhostCoords to include ListIndex now works.
With ReaperCoords, had to replace each coordinate entirely, as it was just a vector at first, and is now a class.
GhostCoords was easier, as it was already a class, so I just had to match up the coords and add in the ListIndex accurately for each coordinate.
Still a lot of debug code in here, but I figure next step is to submit it to the mod testers to see if it works, then post the mod update online.